Do not reset Average; fix types for avg and stddev
Average will roll over after 1000 samples. Do not reset earlier
to make data a bit more interesting.

+++ src/SettingsStorage.h
@@ -39,8 +39,8 @@ public:
 
   unsigned long pidLoopMin = 0;
   unsigned long pidLoopMax = 0;
-  unsigned long pidLoopAvg = 0;
-  unsigned long pidLoopStdDev = 0;
+  float pidLoopAvg = 0;
+  float pidLoopStdDev = 0;
 
   struct PidSettings {
     double desiredTemperature;
